+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - New #191: Add `optionPersistentIds` field to `PollAnswer` type.
 - New #191: Add `PollOptionAdded` and `PollOptionDeleted` type.
 - New #191: Add `pollOptionAdded` and `pollOptionDeleted` fields to `Message` type.
+- New #191: Add `pollOptionId` field to `ReplyParameters` type.
 
 ## 0.17 March 1, 2026